backend/services/auth_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
pyjwt==2.8.1
bcrypt==4.1.1
httpx==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.1
supabase==2.3.5
sqlalchemy==2.0.23
asyncpg==0.29.0
